HA: trying fixing problem with indexes and writeQuorum=1

+++ distributed/src/test/java/com/orientechnologies/orient/server/distributed/AbstractHARemoveNode.java
@@ -42,7 +42,7 @@ public abstract class AbstractHARemoveNode extends AbstractServerClusterTxTest {
     waitFor(2, new OCallable<Boolean, ODatabaseDocumentTx>() {
       @Override
       public Boolean call(ODatabaseDocumentTx db) {
-        final int node2Expected = lastNodeIsUp.get() ? expected : expected - (count * writerCount * (serverInstance.size() - 1));
+        final long node2Expected = lastNodeIsUp.get() ? expected : expected - (count * writerCount * (serverInstance.size() - 1));
 
         final boolean ok = db.countClass("Person") >= node2Expected;
         if (!ok)
